requirements.txt
pytest==8.1.1
pytest-html==4.1.1
requests==2.31.0
pyjwt==2.8.0
flask==3.0.3
python-dotenv==1.0.1
